done: 102, 104; comments: error was an overflow error which turned the int to a negative number. Really awesome to see it in action, but still annoying to get affected by it

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,11 +8,11 @@
  */
 void fibonacci_50(void)
 {
-	int fib_num = 0;
+	long int fib_num = 0;
 
-	int prevnum_a = 1;
+	long int prevnum_a = 1;
 
-	int prevnum_b = 0;
+	long int prevnum_b = 0;
 
 	int i = 0;
